+++ core/src/it/units/sdm/project/game/FreedomPointsCounter.java
@@ -5,7 +5,7 @@ import it.units.sdm.project.enums.Direction;
 import it.units.sdm.project.board.Position;
 import it.units.sdm.project.board.Stone;
 import it.units.sdm.project.exceptions.InvalidPositionException;
-import it.units.sdm.project.interfaces.Board;
+import it.units.sdm.project.board.Board;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
