+++ core/src/it/units/sdm/project/game/gui/FreedomGame.java
@@ -9,7 +9,7 @@ import it.units.sdm.project.board.MapBoard;
 import it.units.sdm.project.enums.GameStatus;
 import it.units.sdm.project.game.Move;
 import it.units.sdm.project.game.Player;
-import it.units.sdm.project.interfaces.Board;
+import it.units.sdm.project.board.Board;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.LinkedList;
