+++ core/test/java/FreedomLineTests.java
@@ -2,7 +2,7 @@
 import it.units.sdm.project.board.Position;
 import it.units.sdm.project.board.Stone;
 import it.units.sdm.project.game.FreedomLine;
-import it.units.sdm.project.interfaces.Board;
+import it.units.sdm.project.board.Board;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
