+++ terminal/src/main/java/it/units/sdm/project/TerminalLauncher.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.Color;
 import it.units.sdm.project.board.MapBoard;
 import it.units.sdm.project.board.Stone;
 import it.units.sdm.project.game.Player;
-import it.units.sdm.project.interfaces.Board;
+import it.units.sdm.project.board.Board;
 
 public class TerminalLauncher {
     public static void main(String[] args) {
